feat: update stock in database after order creation

+++ server/src/main/java/com/digitazon/poscoffee/models/OrderLine.java
@@ -26,7 +26,7 @@ public class OrderLine {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
-  private Short count;
+  private Integer count;
   private Float price;
 
   @OneToOne
