+++ server/src/main/java/com/digitazon/poscoffee/models/helpers/config/ConfigOrderLine.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ConfigOrderLine {
 
-  private Short count;
+  private Integer count;
   private Long product;
   private Long variant;
